Load spirv capability list from spec
