--- external/spirv/spirv-capabilities.txt
@@ -1,238 +1,0 @@
-Matrix
-Shader
-Geometry
-Tessellation
-Addresses
-Linkage
-Kernel
-Vector16
-Float16Buffer
-Float16
-Float64
-Int64
-Int64Atomics
-ImageBasic
-ImageReadWrite
-ImageMipmap
-Pipes
-Groups
-DeviceEnqueue
-LiteralSampler
-AtomicStorage
-Int16
-TessellationPointSize
-GeometryPointSize
-ImageGatherExtended
-StorageImageMultisample
-UniformBufferArrayDynamicIndexing
-SampledImageArrayDynamicIndexing
-StorageBufferArrayDynamicIndexing
-StorageImageArrayDynamicIndexing
-ClipDistance
-CullDistance
-ImageCubeArray
-SampleRateShading
-ImageRect
-SampledRect
-GenericPointer
-Int8
-InputAttachment
-SparseResidency
-MinLod
-Sampled1D
-Image1D
-SampledCubeArray
-SampledBuffer
-ImageBuffer
-ImageMSArray
-StorageImageExtendedFormats
-ImageQuery
-DerivativeControl
-InterpolationFunction
-TransformFeedback
-GeometryStreams
-StorageImageReadWithoutFormat
-StorageImageWriteWithoutFormat
-MultiViewport
-SubgroupDispatch
-NamedBarrier
-PipeStorage
-GroupNonUniform
-GroupNonUniformVote
-GroupNonUniformArithmetic
-GroupNonUniformBallot
-GroupNonUniformShuffle
-GroupNonUniformShuffleRelative
-GroupNonUniformClustered
-GroupNonUniformQuad
-ShaderLayer
-ShaderViewportIndex
-UniformDecoration
-CoreBuiltinsARM
-TileImageColorReadAccessEXT
-TileImageDepthReadAccessEXT
-TileImageStencilReadAccessEXT
-FragmentShadingRateKHR
-SubgroupBallotKHR
-DrawParameters
-WorkgroupMemoryExplicitLayoutKHR
-WorkgroupMemoryExplicitLayout8BitAccessKHR
-WorkgroupMemoryExplicitLayout16BitAccessKHR
-SubgroupVoteKHR
-StorageBuffer16BitAccess
-StorageUniformBufferBlock16
-StorageUniform16
-UniformAndStorageBuffer16BitAccess
-StoragePushConstant16
-StorageInputOutput16
-DeviceGroup
-MultiView
-VariablePointersStorageBuffer
-VariablePointers
-AtomicStorageOps
-SampleMaskPostDepthCoverage
-StorageBuffer8BitAccess
-UniformAndStorageBuffer8BitAccess
-StoragePushConstant8
-DenormPreserve
-DenormFlushToZero
-SignedZeroInfNanPreserve
-RoundingModeRTE
-RoundingModeRTZ
-RayQueryProvisionalKHR
-RayQueryKHR
-RayTraversalPrimitiveCullingKHR
-RayTracingKHR
-Float16ImageAMD
-ImageGatherBiasLodAMD
-FragmentMaskAMD
-StencilExportEXT
-ImageReadWriteLodAMD
-Int64ImageEXT
-ShaderClockKHR
-SampleMaskOverrideCoverageNV
-GeometryShaderPassthroughNV
-ShaderViewportIndexLayerEXT
-ShaderViewportIndexLayerNV
-ShaderViewportMaskNV
-ShaderStereoViewNV
-PerViewAttributesNV
-FragmentFullyCoveredEXT
-MeshShadingNV
-ImageFootprintNV
-MeshShadingEXT
-FragmentBarycentricKHR
-FragmentBarycentricNV
-ComputeDerivativeGroupQuadsNV
-FragmentDensityEXT
-ShadingRateNV
-GroupNonUniformPartitionedNV
-ShaderNonUniform
-ShaderNonUniformEXT
-RuntimeDescriptorArray
-RuntimeDescriptorArrayEXT
-InputAttachmentArrayDynamicIndexing
-InputAttachmentArrayDynamicIndexingEXT
-UniformTexelBufferArrayDynamicIndexing
-UniformTexelBufferArrayDynamicIndexingEXT
-StorageTexelBufferArrayDynamicIndexing
-StorageTexelBufferArrayDynamicIndexingEXT
-UniformBufferArrayNonUniformIndexing
-UniformBufferArrayNonUniformIndexingEXT
-SampledImageArrayNonUniformIndexing
-SampledImageArrayNonUniformIndexingEXT
-StorageBufferArrayNonUniformIndexing
-StorageBufferArrayNonUniformIndexingEXT
-StorageImageArrayNonUniformIndexing
-StorageImageArrayNonUniformIndexingEXT
-InputAttachmentArrayNonUniformIndexing
-InputAttachmentArrayNonUniformIndexingEXT
-UniformTexelBufferArrayNonUniformIndexing
-UniformTexelBufferArrayNonUniformIndexingEXT
-StorageTexelBufferArrayNonUniformIndexing
-StorageTexelBufferArrayNonUniformIndexingEXT
-RayTracingPositionFetchKHR
-RayTracingNV
-RayTracingMotionBlurNV
-VulkanMemoryModel
-VulkanMemoryModelKHR
-VulkanMemoryModelDeviceScope
-VulkanMemoryModelDeviceScopeKHR
-PhysicalStorageBufferAddresses
-PhysicalStorageBufferAddressesEXT
-ComputeDerivativeGroupLinearNV
-RayTracingProvisionalKHR
-CooperativeMatrixNV
-FragmentShaderSampleInterlockEXT
-FragmentShaderShadingRateInterlockEXT
-ShaderSMBuiltinsNV
-FragmentShaderPixelInterlockEXT
-DemoteToHelperInvocation
-DemoteToHelperInvocationEXT
-RayTracingOpacityMicromapEXT
-ShaderInvocationReorderNV
-BindlessTextureNV
-RayQueryPositionFetchKHR
-SubgroupShuffleINTEL
-SubgroupBufferBlockIOINTEL
-SubgroupImageBlockIOINTEL
-SubgroupImageMediaBlockIOINTEL
-RoundToInfinityINTEL
-FloatingPointModeINTEL
-IntegerFunctions2INTEL
-FunctionPointersINTEL
-IndirectReferencesINTEL
-AsmINTEL
-AtomicFloat32MinMaxEXT
-AtomicFloat64MinMaxEXT
-AtomicFloat16MinMaxEXT
-VectorComputeINTEL
-VectorAnyINTEL
-ExpectAssumeKHR
-SubgroupAvcMotionEstimationINTEL
-SubgroupAvcMotionEstimationIntraINTEL
-SubgroupAvcMotionEstimationChromaINTEL
-VariableLengthArrayINTEL
-FunctionFloatControlINTEL
-FPGAMemoryAttributesINTEL
-FPFastMathModeINTEL
-ArbitraryPrecisionIntegersINTEL
-ArbitraryPrecisionFloatingPointINTEL
-UnstructuredLoopControlsINTEL
-FPGALoopControlsINTEL
-KernelAttributesINTEL
-FPGAKernelAttributesINTEL
-FPGAMemoryAccessesINTEL
-FPGAClusterAttributesINTEL
-LoopFuseINTEL
-FPGADSPControlINTEL
-MemoryAccessAliasingINTEL
-FPGAInvocationPipeliningAttributesINTEL
-FPGABufferLocationINTEL
-ArbitraryPrecisionFixedPointINTEL
-USMStorageClassesINTEL
-RuntimeAlignedAttributeINTEL
-IOPipesINTEL
-BlockingPipesINTEL
-FPGARegINTEL
-DotProductInputAll
-DotProductInputAllKHR
-DotProductInput4x8Bit
-DotProductInput4x8BitKHR
-DotProductInput4x8BitPacked
-DotProductInput4x8BitPackedKHR
-DotProduct
-DotProductKHR
-RayCullMaskKHR
-BitInstructions
-GroupNonUniformRotateKHR
-AtomicFloat32AddEXT
-AtomicFloat64AddEXT
-LongConstantCompositeINTEL
-OptNoneINTEL
-AtomicFloat16AddEXT
-DebugInfoModuleINTEL
-SplitBarrierINTEL
-FPGAArgumentInterfacesINTEL
-GroupUniformArithmeticKHR
-Max
