+++ premake5.lua
@@ -1444,12 +1444,6 @@ generatorProject("generate-lookup-tables")
             prefix = "GLSLstd450",
             type = "GLSLstd450"
         },
-        {
-            json = "external/spirv/spirv-capabilities.txt",
-            header = "spirv/unified1/spirv.h",
-            prefix = "SpvCapability",
-            type = "SpvCapability"
-        },
     }
     for _, t in pairs(tables) do
         files {t.json}
@@ -1640,7 +1634,6 @@ standardProject("slang", "source/slang")
     -- Similarly for any generated lookup tables
     files {
         "source/slang/slang-lookup-glslstd450.cpp",
-        "source/slang/slang-lookup-spvcapability.cpp",
     }
 
     --
