+++ source/compiler-core/slang-spirv-core-grammar.cpp
@@ -1,6 +1,7 @@
 #include "slang-spirv-core-grammar.h"
 
 #include "../core/slang-rtti-util.h"
+#include "../core/slang-string-util.h"
 #include "slang-json-native.h"
 #include "slang-core-diagnostics.h"
 
@@ -33,6 +34,7 @@ SLANG_MAKE_STRUCT_RTTI_INFO(
     Operand,
     SLANG_RTTI_FIELD(kind),
     SLANG_OPTIONAL_RTTI_FIELD(quantifier)
+    //SLANG_RTTI_FIELD(name),
 );
 
 struct Instruction
@@ -52,9 +54,42 @@ SLANG_MAKE_STRUCT_RTTI_INFO(
     SLANG_OPTIONAL_RTTI_FIELD(operands)
 );
 
+struct Enumerant
+{
+    UnownedStringSlice enumerant;
+    JSONValue value;
+    List<UnownedStringSlice> capabilities;
+    // List<Operand> parameters;
+    // UnownedStringSlice version;
+    // UnownedStringSlice lastVersion;
+    // List<UnownedStringSlice> extensions;
+};
+SLANG_MAKE_STRUCT_RTTI_INFO(
+    Enumerant,
+    SLANG_RTTI_FIELD(enumerant),
+    SLANG_RTTI_FIELD(value),
+    SLANG_OPTIONAL_RTTI_FIELD(capabilities),
+    // SLANG_OPTIONAL_RTTI_FIELD(parameters),
+    // SLANG_OPTIONAL_RTTI_FIELD(version),
+    // SLANG_OPTIONAL_RTTI_FIELD(lastVersion),
+    // SLANG_OPTIONAL_RTTI_FIELD(extensions)
+);
+
+struct OperandKind
+{
+    UnownedStringSlice category;
+    UnownedStringSlice kind;
+    List<Enumerant> enumerants;
+};
+SLANG_MAKE_STRUCT_RTTI_INFO(
+    OperandKind,
+    SLANG_RTTI_FIELD(category),
+    SLANG_RTTI_FIELD(kind),
+    SLANG_OPTIONAL_RTTI_FIELD(enumerants)
+);
+
 struct SPIRVSpec
 {
-    // We don't currently use these
     // List<UnownedStringSlice> copyright;
     // UnownedStringSlice magic_number;
     // UInt32 major_version;
@@ -62,6 +97,7 @@ struct SPIRVSpec
     // UInt32 revision;
     List<InstructionPrintingClass> instruction_printing_class;
     List<Instruction> instructions;
+    List<OperandKind> operand_kinds;
 };
 SLANG_MAKE_STRUCT_RTTI_INFO(
     SPIRVSpec,
@@ -71,7 +107,8 @@ SLANG_MAKE_STRUCT_RTTI_INFO(
     // SLANG_RTTI_FIELD(minor_version)
     // SLANG_RTTI_FIELD(revision)
     SLANG_RTTI_FIELD(instruction_printing_class),
-    SLANG_RTTI_FIELD(instructions)
+    SLANG_RTTI_FIELD(instructions),
+    SLANG_RTTI_FIELD(operand_kinds)
 );
 
 //
@@ -107,6 +144,40 @@ RefPtr<SPIRVCoreGrammarInfo> loadSPIRVCoreGrammarInfo(SourceView& source, Diagno
     res->spvOps.dict.reserve(spec.instructions.getCount());
     for(const auto& i : spec.instructions)
         res->spvOps.dict.add(i.opname, SpvOp(i.opcode));
+    for(const auto& k : spec.operand_kinds)
+    {
+        if(k.kind == "Capability")
+        {
+            res->spvCapabilities.dict.reserve(k.enumerants.getCapacity());
+            for(const auto& e : k.enumerants)
+            {
+                SpvCapability valueInt;
+                switch(e.value.getKind())
+                {
+                    case JSONValue::Kind::Integer:
+                    {
+                        // TODO: Range check here?
+                        valueInt = SpvCapability(container.asInteger(e.value));
+                        break;
+                    }
+                    case JSONValue::Kind::String:
+                    {
+                        Int i = 0;
+                        const auto str = container.getString(e.value);
+                        if(SLANG_FAILED(StringUtil::parseInt(str, i)))
+                            sink.diagnose(e.value.loc, MiscDiagnostics::spirvCoreGrammarJSONParseFailure);
+                        // TODO: Range check here?
+                        valueInt = SpvCapability(i);
+                        break;
+                     }
+                     default:
+                        sink.diagnose(e.value.loc, MiscDiagnostics::spirvCoreGrammarJSONParseFailure);
+                }
+                res->spvCapabilities.dict.add(e.enumerant, valueInt);
+            }
+            break;
+        }
+    }
     return res;
 }
 }
