--- source/slang/slang-lookup-spvcapability.cpp
@@ -1,295 +1,0 @@
-// Hash function for SpvCapability
-//
-// This file was thoughtfully generated by a machine,
-// don't even think about modifying it yourself!
-//
-
-#include "../core/slang-common.h"
-#include "../core/slang-string.h"
-#include "spirv/unified1/spirv.h"
-
-
-namespace Slang
-{
-
-bool lookupSpvCapability(const UnownedStringSlice& str, SpvCapability& value)
-{
-    static const unsigned tableSalt[238] = {
-        0, 5, 1, 13, 1, 1, 1, 0, 2, 0, 1, 1, 0, 1, 6, 2,
-        6, 0, 0, 7, 2, 0, 0, 7, 6, 2, 2, 5, 1, 0, 1, 1,
-        0, 0, 0, 0, 7, 0, 0, 2, 0, 0, 2, 10, 3, 0, 0, 2,
-        7, 1, 0, 4, 5, 3, 0, 0, 3, 3, 0, 0, 1, 1, 0, 2,
-        0, 3, 1, 2, 0, 3, 5, 0, 10, 4, 2, 0, 0, 0, 0, 2,
-        1, 12, 0, 12, 6, 0, 13, 7, 0, 0, 0, 2, 0, 0, 0, 3,
-        8, 5, 13, 0, 1, 0, 4, 0, 5, 0, 1, 5, 1, 0, 1, 0,
-        7, 9, 5, 2, 0, 3, 9, 0, 0, 0, 14, 9, 2, 6, 0, 1,
-        0, 1, 2, 3, 4, 2, 2, 8, 7, 1, 21, 2, 0, 2, 0, 1,
-        13, 15, 0, 11, 1, 0, 1, 0, 0, 1, 2, 0, 6, 8, 10, 0,
-        2, 0, 1, 2, 10, 0, 0, 0, 1, 3, 3, 25, 0, 0, 0, 0,
-        0, 1, 15, 88, 0, 18, 1, 1, 3, 0, 0, 0, 0, 1, 0, 7,
-        0, 0, 4, 8, 13, 0, 0, 33, 3, 60, 0, 2, 4, 17, 0, 1,
-        0, 0, 2, 53, 0, 0, 0, 0, 1, 76, 0, 0, 14, 2, 219, 0,
-        6, 49, 0, 0, 0, 2, 1, 0, 6, 630, 0, 0, 2, 2
-    };
-
-    using KV = std::pair<const char*, SpvCapability>;
-
-    static const KV words[238] =
-    {
-        {"ShaderInvocationReorderNV", SpvCapabilityShaderInvocationReorderNV},
-        {"RayQueryProvisionalKHR", SpvCapabilityRayQueryProvisionalKHR},
-        {"Max", SpvCapabilityMax},
-        {"ShaderLayer", SpvCapabilityShaderLayer},
-        {"ImageBuffer", SpvCapabilityImageBuffer},
-        {"UnstructuredLoopControlsINTEL", SpvCapabilityUnstructuredLoopControlsINTEL},
-        {"GeometryStreams", SpvCapabilityGeometryStreams},
-        {"AtomicFloat16AddEXT", SpvCapabilityAtomicFloat16AddEXT},
-        {"UniformDecoration", SpvCapabilityUniformDecoration},
-        {"StorageTexelBufferArrayDynamicIndexing", SpvCapabilityStorageTexelBufferArrayDynamicIndexing},
-        {"OptNoneINTEL", SpvCapabilityOptNoneINTEL},
-        {"FragmentDensityEXT", SpvCapabilityFragmentDensityEXT},
-        {"IOPipesINTEL", SpvCapabilityIOPipesINTEL},
-        {"ImageMSArray", SpvCapabilityImageMSArray},
-        {"ShaderSMBuiltinsNV", SpvCapabilityShaderSMBuiltinsNV},
-        {"SubgroupImageBlockIOINTEL", SpvCapabilitySubgroupImageBlockIOINTEL},
-        {"LiteralSampler", SpvCapabilityLiteralSampler},
-        {"KernelAttributesINTEL", SpvCapabilityKernelAttributesINTEL},
-        {"AtomicFloat64AddEXT", SpvCapabilityAtomicFloat64AddEXT},
-        {"RayTraversalPrimitiveCullingKHR", SpvCapabilityRayTraversalPrimitiveCullingKHR},
-        {"Shader", SpvCapabilityShader},
-        {"GroupUniformArithmeticKHR", SpvCapabilityGroupUniformArithmeticKHR},
-        {"DenormPreserve", SpvCapabilityDenormPreserve},
-        {"IndirectReferencesINTEL", SpvCapabilityIndirectReferencesINTEL},
-        {"SubgroupDispatch", SpvCapabilitySubgroupDispatch},
-        {"SubgroupBallotKHR", SpvCapabilitySubgroupBallotKHR},
-        {"RuntimeDescriptorArrayEXT", SpvCapabilityRuntimeDescriptorArrayEXT},
-        {"Kernel", SpvCapabilityKernel},
-        {"VulkanMemoryModelDeviceScopeKHR", SpvCapabilityVulkanMemoryModelDeviceScopeKHR},
-        {"StoragePushConstant16", SpvCapabilityStoragePushConstant16},
-        {"InputAttachmentArrayNonUniformIndexing", SpvCapabilityInputAttachmentArrayNonUniformIndexing},
-        {"BlockingPipesINTEL", SpvCapabilityBlockingPipesINTEL},
-        {"SplitBarrierINTEL", SpvCapabilitySplitBarrierINTEL},
-        {"FPGAMemoryAttributesINTEL", SpvCapabilityFPGAMemoryAttributesINTEL},
-        {"UniformBufferArrayDynamicIndexing", SpvCapabilityUniformBufferArrayDynamicIndexing},
-        {"AtomicFloat32AddEXT", SpvCapabilityAtomicFloat32AddEXT},
-        {"Image1D", SpvCapabilityImage1D},
-        {"VariablePointers", SpvCapabilityVariablePointers},
-        {"FPGAInvocationPipeliningAttributesINTEL", SpvCapabilityFPGAInvocationPipeliningAttributesINTEL},
-        {"ImageCubeArray", SpvCapabilityImageCubeArray},
-        {"StorageImageArrayDynamicIndexing", SpvCapabilityStorageImageArrayDynamicIndexing},
-        {"VulkanMemoryModelKHR", SpvCapabilityVulkanMemoryModelKHR},
-        {"FragmentShadingRateKHR", SpvCapabilityFragmentShadingRateKHR},
-        {"RayTracingProvisionalKHR", SpvCapabilityRayTracingProvisionalKHR},
-        {"Vector16", SpvCapabilityVector16},
-        {"DotProductInputAll", SpvCapabilityDotProductInputAll},
-        {"ArbitraryPrecisionIntegersINTEL", SpvCapabilityArbitraryPrecisionIntegersINTEL},
-        {"VectorComputeINTEL", SpvCapabilityVectorComputeINTEL},
-        {"AtomicFloat64MinMaxEXT", SpvCapabilityAtomicFloat64MinMaxEXT},
-        {"VulkanMemoryModelDeviceScope", SpvCapabilityVulkanMemoryModelDeviceScope},
-        {"FPGALoopControlsINTEL", SpvCapabilityFPGALoopControlsINTEL},
-        {"Float16ImageAMD", SpvCapabilityFloat16ImageAMD},
-        {"Int64ImageEXT", SpvCapabilityInt64ImageEXT},
-        {"Int64Atomics", SpvCapabilityInt64Atomics},
-        {"DotProductInput4x8Bit", SpvCapabilityDotProductInput4x8Bit},
-        {"Addresses", SpvCapabilityAddresses},
-        {"DrawParameters", SpvCapabilityDrawParameters},
-        {"RoundToInfinityINTEL", SpvCapabilityRoundToInfinityINTEL},
-        {"SubgroupAvcMotionEstimationINTEL", SpvCapabilitySubgroupAvcMotionEstimationINTEL},
-        {"FPGAArgumentInterfacesINTEL", SpvCapabilityFPGAArgumentInterfacesINTEL},
-        {"StorageImageArrayNonUniformIndexing", SpvCapabilityStorageImageArrayNonUniformIndexing},
-        {"LongConstantCompositeINTEL", SpvCapabilityLongConstantCompositeINTEL},
-        {"WorkgroupMemoryExplicitLayout16BitAccessKHR", SpvCapabilityWorkgroupMemoryExplicitLayout16BitAccessKHR},
-        {"StorageBuffer16BitAccess", SpvCapabilityStorageBuffer16BitAccess},
-        {"StorageTexelBufferArrayDynamicIndexingEXT", SpvCapabilityStorageTexelBufferArrayDynamicIndexingEXT},
-        {"USMStorageClassesINTEL", SpvCapabilityUSMStorageClassesINTEL},
-        {"FPGARegINTEL", SpvCapabilityFPGARegINTEL},
-        {"Matrix", SpvCapabilityMatrix},
-        {"StorageBuffer8BitAccess", SpvCapabilityStorageBuffer8BitAccess},
-        {"SampledRect", SpvCapabilitySampledRect},
-        {"SparseResidency", SpvCapabilitySparseResidency},
-        {"SubgroupShuffleINTEL", SpvCapabilitySubgroupShuffleINTEL},
-        {"VulkanMemoryModel", SpvCapabilityVulkanMemoryModel},
-        {"ImageQuery", SpvCapabilityImageQuery},
-        {"RayCullMaskKHR", SpvCapabilityRayCullMaskKHR},
-        {"SampleRateShading", SpvCapabilitySampleRateShading},
-        {"ShaderViewportIndexLayerEXT", SpvCapabilityShaderViewportIndexLayerEXT},
-        {"UniformAndStorageBuffer8BitAccess", SpvCapabilityUniformAndStorageBuffer8BitAccess},
-        {"GroupNonUniformArithmetic", SpvCapabilityGroupNonUniformArithmetic},
-        {"StorageTexelBufferArrayNonUniformIndexingEXT", SpvCapabilityStorageTexelBufferArrayNonUniformIndexingEXT},
-        {"SubgroupAvcMotionEstimationChromaINTEL", SpvCapabilitySubgroupAvcMotionEstimationChromaINTEL},
-        {"Float16Buffer", SpvCapabilityFloat16Buffer},
-        {"UniformAndStorageBuffer16BitAccess", SpvCapabilityUniformAndStorageBuffer16BitAccess},
-        {"ClipDistance", SpvCapabilityClipDistance},
-        {"RuntimeDescriptorArray", SpvCapabilityRuntimeDescriptorArray},
-        {"RayQueryKHR", SpvCapabilityRayQueryKHR},
-        {"DebugInfoModuleINTEL", SpvCapabilityDebugInfoModuleINTEL},
-        {"VariablePointersStorageBuffer", SpvCapabilityVariablePointersStorageBuffer},
-        {"WorkgroupMemoryExplicitLayout8BitAccessKHR", SpvCapabilityWorkgroupMemoryExplicitLayout8BitAccessKHR},
-        {"FPGAClusterAttributesINTEL", SpvCapabilityFPGAClusterAttributesINTEL},
-        {"Int64", SpvCapabilityInt64},
-        {"Sampled1D", SpvCapabilitySampled1D},
-        {"InterpolationFunction", SpvCapabilityInterpolationFunction},
-        {"AtomicStorageOps", SpvCapabilityAtomicStorageOps},
-        {"ShaderViewportIndex", SpvCapabilityShaderViewportIndex},
-        {"FloatingPointModeINTEL", SpvCapabilityFloatingPointModeINTEL},
-        {"FPFastMathModeINTEL", SpvCapabilityFPFastMathModeINTEL},
-        {"SampleMaskPostDepthCoverage", SpvCapabilitySampleMaskPostDepthCoverage},
-        {"UniformBufferArrayNonUniformIndexingEXT", SpvCapabilityUniformBufferArrayNonUniformIndexingEXT},
-        {"WorkgroupMemoryExplicitLayoutKHR", SpvCapabilityWorkgroupMemoryExplicitLayoutKHR},
-        {"GeometryPointSize", SpvCapabilityGeometryPointSize},
-        {"ArbitraryPrecisionFloatingPointINTEL", SpvCapabilityArbitraryPrecisionFloatingPointINTEL},
-        {"DeviceEnqueue", SpvCapabilityDeviceEnqueue},
-        {"LoopFuseINTEL", SpvCapabilityLoopFuseINTEL},
-        {"InputAttachment", SpvCapabilityInputAttachment},
-        {"MeshShadingEXT", SpvCapabilityMeshShadingEXT},
-        {"FPGAKernelAttributesINTEL", SpvCapabilityFPGAKernelAttributesINTEL},
-        {"ShaderViewportIndexLayerNV", SpvCapabilityShaderViewportIndexLayerNV},
-        {"ExpectAssumeKHR", SpvCapabilityExpectAssumeKHR},
-        {"FragmentBarycentricNV", SpvCapabilityFragmentBarycentricNV},
-        {"Int16", SpvCapabilityInt16},
-        {"UniformTexelBufferArrayDynamicIndexing", SpvCapabilityUniformTexelBufferArrayDynamicIndexing},
-        {"ImageBasic", SpvCapabilityImageBasic},
-        {"IntegerFunctions2INTEL", SpvCapabilityIntegerFunctions2INTEL},
-        {"GroupNonUniformRotateKHR", SpvCapabilityGroupNonUniformRotateKHR},
-        {"UniformTexelBufferArrayNonUniformIndexing", SpvCapabilityUniformTexelBufferArrayNonUniformIndexing},
-        {"DotProductKHR", SpvCapabilityDotProductKHR},
-        {"DenormFlushToZero", SpvCapabilityDenormFlushToZero},
-        {"ShaderNonUniform", SpvCapabilityShaderNonUniform},
-        {"StorageUniform16", SpvCapabilityStorageUniform16},
-        {"GroupNonUniformBallot", SpvCapabilityGroupNonUniformBallot},
-        {"StoragePushConstant8", SpvCapabilityStoragePushConstant8},
-        {"SampledBuffer", SpvCapabilitySampledBuffer},
-        {"FunctionPointersINTEL", SpvCapabilityFunctionPointersINTEL},
-        {"CooperativeMatrixNV", SpvCapabilityCooperativeMatrixNV},
-        {"PipeStorage", SpvCapabilityPipeStorage},
-        {"VectorAnyINTEL", SpvCapabilityVectorAnyINTEL},
-        {"SampledImageArrayDynamicIndexing", SpvCapabilitySampledImageArrayDynamicIndexing},
-        {"TessellationPointSize", SpvCapabilityTessellationPointSize},
-        {"StorageImageWriteWithoutFormat", SpvCapabilityStorageImageWriteWithoutFormat},
-        {"FragmentBarycentricKHR", SpvCapabilityFragmentBarycentricKHR},
-        {"Geometry", SpvCapabilityGeometry},
-        {"ComputeDerivativeGroupQuadsNV", SpvCapabilityComputeDerivativeGroupQuadsNV},
-        {"DotProductInputAllKHR", SpvCapabilityDotProductInputAllKHR},
-        {"RayTracingMotionBlurNV", SpvCapabilityRayTracingMotionBlurNV},
-        {"RoundingModeRTZ", SpvCapabilityRoundingModeRTZ},
-        {"DotProductInput4x8BitPacked", SpvCapabilityDotProductInput4x8BitPacked},
-        {"ImageReadWrite", SpvCapabilityImageReadWrite},
-        {"SubgroupBufferBlockIOINTEL", SpvCapabilitySubgroupBufferBlockIOINTEL},
-        {"FPGADSPControlINTEL", SpvCapabilityFPGADSPControlINTEL},
-        {"PhysicalStorageBufferAddressesEXT", SpvCapabilityPhysicalStorageBufferAddressesEXT},
-        {"MultiView", SpvCapabilityMultiView},
-        {"FPGAMemoryAccessesINTEL", SpvCapabilityFPGAMemoryAccessesINTEL},
-        {"GroupNonUniform", SpvCapabilityGroupNonUniform},
-        {"MultiViewport", SpvCapabilityMultiViewport},
-        {"DemoteToHelperInvocationEXT", SpvCapabilityDemoteToHelperInvocationEXT},
-        {"DeviceGroup", SpvCapabilityDeviceGroup},
-        {"CoreBuiltinsARM", SpvCapabilityCoreBuiltinsARM},
-        {"StorageBufferArrayDynamicIndexing", SpvCapabilityStorageBufferArrayDynamicIndexing},
-        {"FragmentMaskAMD", SpvCapabilityFragmentMaskAMD},
-        {"Float64", SpvCapabilityFloat64},
-        {"BitInstructions", SpvCapabilityBitInstructions},
-        {"MemoryAccessAliasingINTEL", SpvCapabilityMemoryAccessAliasingINTEL},
-        {"GenericPointer", SpvCapabilityGenericPointer},
-        {"UniformBufferArrayNonUniformIndexing", SpvCapabilityUniformBufferArrayNonUniformIndexing},
-        {"FragmentShaderSampleInterlockEXT", SpvCapabilityFragmentShaderSampleInterlockEXT},
-        {"InputAttachmentArrayNonUniformIndexingEXT", SpvCapabilityInputAttachmentArrayNonUniformIndexingEXT},
-        {"StorageTexelBufferArrayNonUniformIndexing", SpvCapabilityStorageTexelBufferArrayNonUniformIndexing},
-        {"GroupNonUniformClustered", SpvCapabilityGroupNonUniformClustered},
-        {"Groups", SpvCapabilityGroups},
-        {"TileImageDepthReadAccessEXT", SpvCapabilityTileImageDepthReadAccessEXT},
-        {"VariableLengthArrayINTEL", SpvCapabilityVariableLengthArrayINTEL},
-        {"ArbitraryPrecisionFixedPointINTEL", SpvCapabilityArbitraryPrecisionFixedPointINTEL},
-        {"GroupNonUniformPartitionedNV", SpvCapabilityGroupNonUniformPartitionedNV},
-        {"DerivativeControl", SpvCapabilityDerivativeControl},
-        {"SignedZeroInfNanPreserve", SpvCapabilitySignedZeroInfNanPreserve},
-        {"InputAttachmentArrayDynamicIndexingEXT", SpvCapabilityInputAttachmentArrayDynamicIndexingEXT},
-        {"GroupNonUniformShuffle", SpvCapabilityGroupNonUniformShuffle},
-        {"TileImageStencilReadAccessEXT", SpvCapabilityTileImageStencilReadAccessEXT},
-        {"ShaderStereoViewNV", SpvCapabilityShaderStereoViewNV},
-        {"StorageImageMultisample", SpvCapabilityStorageImageMultisample},
-        {"FragmentShaderShadingRateInterlockEXT", SpvCapabilityFragmentShaderShadingRateInterlockEXT},
-        {"SubgroupImageMediaBlockIOINTEL", SpvCapabilitySubgroupImageMediaBlockIOINTEL},
-        {"ShadingRateNV", SpvCapabilityShadingRateNV},
-        {"ImageGatherBiasLodAMD", SpvCapabilityImageGatherBiasLodAMD},
-        {"BindlessTextureNV", SpvCapabilityBindlessTextureNV},
-        {"SampledCubeArray", SpvCapabilitySampledCubeArray},
-        {"ImageReadWriteLodAMD", SpvCapabilityImageReadWriteLodAMD},
-        {"InputAttachmentArrayDynamicIndexing", SpvCapabilityInputAttachmentArrayDynamicIndexing},
-        {"AtomicFloat32MinMaxEXT", SpvCapabilityAtomicFloat32MinMaxEXT},
-        {"DotProductInput4x8BitKHR", SpvCapabilityDotProductInput4x8BitKHR},
-        {"AtomicStorage", SpvCapabilityAtomicStorage},
-        {"PerViewAttributesNV", SpvCapabilityPerViewAttributesNV},
-        {"RayTracingKHR", SpvCapabilityRayTracingKHR},
-        {"StorageBufferArrayNonUniformIndexing", SpvCapabilityStorageBufferArrayNonUniformIndexing},
-        {"StorageUniformBufferBlock16", SpvCapabilityStorageUniformBufferBlock16},
-        {"StorageImageExtendedFormats", SpvCapabilityStorageImageExtendedFormats},
-        {"GroupNonUniformQuad", SpvCapabilityGroupNonUniformQuad},
-        {"TileImageColorReadAccessEXT", SpvCapabilityTileImageColorReadAccessEXT},
-        {"ImageGatherExtended", SpvCapabilityImageGatherExtended},
-        {"SampledImageArrayNonUniformIndexingEXT", SpvCapabilitySampledImageArrayNonUniformIndexingEXT},
-        {"PhysicalStorageBufferAddresses", SpvCapabilityPhysicalStorageBufferAddresses},
-        {"SubgroupAvcMotionEstimationIntraINTEL", SpvCapabilitySubgroupAvcMotionEstimationIntraINTEL},
-        {"GroupNonUniformShuffleRelative", SpvCapabilityGroupNonUniformShuffleRelative},
-        {"UniformTexelBufferArrayDynamicIndexingEXT", SpvCapabilityUniformTexelBufferArrayDynamicIndexingEXT},
-        {"AsmINTEL", SpvCapabilityAsmINTEL},
-        {"FragmentFullyCoveredEXT", SpvCapabilityFragmentFullyCoveredEXT},
-        {"RayTracingOpacityMicromapEXT", SpvCapabilityRayTracingOpacityMicromapEXT},
-        {"CullDistance", SpvCapabilityCullDistance},
-        {"Pipes", SpvCapabilityPipes},
-        {"MinLod", SpvCapabilityMinLod},
-        {"ShaderNonUniformEXT", SpvCapabilityShaderNonUniformEXT},
-        {"MeshShadingNV", SpvCapabilityMeshShadingNV},
-        {"TransformFeedback", SpvCapabilityTransformFeedback},
-        {"Tessellation", SpvCapabilityTessellation},
-        {"Linkage", SpvCapabilityLinkage},
-        {"RayTracingNV", SpvCapabilityRayTracingNV},
-        {"StorageImageArrayNonUniformIndexingEXT", SpvCapabilityStorageImageArrayNonUniformIndexingEXT},
-        {"RayQueryPositionFetchKHR", SpvCapabilityRayQueryPositionFetchKHR},
-        {"StorageInputOutput16", SpvCapabilityStorageInputOutput16},
-        {"ImageRect", SpvCapabilityImageRect},
-        {"StencilExportEXT", SpvCapabilityStencilExportEXT},
-        {"GeometryShaderPassthroughNV", SpvCapabilityGeometryShaderPassthroughNV},
-        {"RuntimeAlignedAttributeINTEL", SpvCapabilityRuntimeAlignedAttributeINTEL},
-        {"UniformTexelBufferArrayNonUniformIndexingEXT", SpvCapabilityUniformTexelBufferArrayNonUniformIndexingEXT},
-        {"ImageFootprintNV", SpvCapabilityImageFootprintNV},
-        {"GroupNonUniformVote", SpvCapabilityGroupNonUniformVote},
-        {"SampledImageArrayNonUniformIndexing", SpvCapabilitySampledImageArrayNonUniformIndexing},
-        {"Int8", SpvCapabilityInt8},
-        {"FragmentShaderPixelInterlockEXT", SpvCapabilityFragmentShaderPixelInterlockEXT},
-        {"ShaderClockKHR", SpvCapabilityShaderClockKHR},
-        {"AtomicFloat16MinMaxEXT", SpvCapabilityAtomicFloat16MinMaxEXT},
-        {"SubgroupVoteKHR", SpvCapabilitySubgroupVoteKHR},
-        {"FunctionFloatControlINTEL", SpvCapabilityFunctionFloatControlINTEL},
-        {"StorageBufferArrayNonUniformIndexingEXT", SpvCapabilityStorageBufferArrayNonUniformIndexingEXT},
-        {"DotProductInput4x8BitPackedKHR", SpvCapabilityDotProductInput4x8BitPackedKHR},
-        {"DemoteToHelperInvocation", SpvCapabilityDemoteToHelperInvocation},
-        {"DotProduct", SpvCapabilityDotProduct},
-        {"NamedBarrier", SpvCapabilityNamedBarrier},
-        {"StorageImageReadWithoutFormat", SpvCapabilityStorageImageReadWithoutFormat},
-        {"ImageMipmap", SpvCapabilityImageMipmap},
-        {"ShaderViewportMaskNV", SpvCapabilityShaderViewportMaskNV},
-        {"Float16", SpvCapabilityFloat16},
-        {"FPGABufferLocationINTEL", SpvCapabilityFPGABufferLocationINTEL},
-        {"RoundingModeRTE", SpvCapabilityRoundingModeRTE},
-        {"SampleMaskOverrideCoverageNV", SpvCapabilitySampleMaskOverrideCoverageNV},
-        {"ComputeDerivativeGroupLinearNV", SpvCapabilityComputeDerivativeGroupLinearNV},
-        {"RayTracingPositionFetchKHR", SpvCapabilityRayTracingPositionFetchKHR},
-    };
-
-    static const auto hash = [](const UnownedStringSlice& str, UInt32 salt){
-        return combineHash(getHashCode(str), getHashCode(salt)) % 238;
-    };
-
-    const auto i = hash(str, tableSalt[hash(str, 0)]);
-    if(str == words[i].first)
-    {
-        value = words[i].second;
-        return true;
-    }
-    else
-    {
-        return false;
-    }
-}
-
-}
