+++ source/slang/slang-spirv-core-grammar-embed.cpp
@@ -798,11 +798,300 @@ bool lookupSpvOp(const UnownedStringSlice& str, SpvOp& value)
     }
 }
 
+bool lookupSpvCapability(const UnownedStringSlice& str, SpvCapability& value)
+{
+    static const unsigned tableSalt[245] = {
+        0, 4, 0, 0, 1, 0, 4, 0, 4, 7, 2, 0, 0, 0, 0, 1,
+        2, 6, 0, 1, 4, 0, 0, 1, 1, 1, 0, 1, 0, 0, 4, 1,
+        1, 1, 0, 0, 3, 0, 1, 1, 0, 0, 7, 1, 1, 2, 6, 7,
+        3, 0, 4, 0, 0, 1, 0, 0, 6, 2, 0, 0, 3, 0, 0, 0,
+        0, 3, 0, 0, 1, 12, 0, 1, 2, 0, 6, 2, 0, 1, 5, 1,
+        5, 1, 0, 6, 12, 0, 5, 9, 2, 11, 0, 0, 3, 2, 1, 0,
+        6, 0, 1, 1, 1, 3, 1, 0, 0, 0, 1, 5, 2, 1, 0, 9,
+        3, 3, 0, 1, 1, 0, 6, 0, 0, 0, 1, 0, 2, 3, 0, 1,
+        2, 2, 6, 4, 2, 0, 3, 2, 5, 7, 5, 3, 2, 2, 0, 1,
+        1, 13, 13, 1, 0, 1, 0, 0, 0, 1, 0, 16, 2, 0, 4, 13,
+        5, 1, 2, 3, 0, 10, 2, 8, 4, 0, 5, 1, 11, 6, 0, 7,
+        18, 10, 0, 55, 4, 12, 0, 0, 0, 1, 5, 0, 13, 13, 0, 2,
+        3, 0, 0, 0, 3, 1, 81, 1, 0, 0, 2, 23, 8, 0, 1, 11,
+        0, 13, 5, 0, 52, 12, 0, 0, 2, 8, 0, 7, 0, 8, 15, 11,
+        10, 72, 125, 0, 0, 36, 16, 1, 318, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 10, 10, 8
+    };
+
+    using KV = std::pair<const char*, SpvCapability>;
+
+    static const KV words[245] =
+    {
+        {"UniformDecoration", SpvCapabilityUniformDecoration},
+        {"ComputeDerivativeGroupLinearNV", SpvCapabilityComputeDerivativeGroupLinearNV},
+        {"InputAttachmentArrayNonUniformIndexing", SpvCapabilityInputAttachmentArrayNonUniformIndexing},
+        {"DenormFlushToZero", SpvCapabilityDenormFlushToZero},
+        {"Shader", SpvCapabilityShader},
+        {"AsmINTEL", SpvCapabilityAsmINTEL},
+        {"ArbitraryPrecisionIntegersINTEL", SpvCapabilityArbitraryPrecisionIntegersINTEL},
+        {"VulkanMemoryModel", SpvCapabilityVulkanMemoryModel},
+        {"SampledCubeArray", SpvCapabilitySampledCubeArray},
+        {"SubgroupDispatch", SpvCapabilitySubgroupDispatch},
+        {"ShaderClockKHR", SpvCapabilityShaderClockKHR},
+        {"Image1D", SpvCapabilityImage1D},
+        {"UniformBufferArrayNonUniformIndexing", SpvCapabilityUniformBufferArrayNonUniformIndexing},
+        {"UniformTexelBufferArrayDynamicIndexing", SpvCapabilityUniformTexelBufferArrayDynamicIndexing},
+        {"VulkanMemoryModelDeviceScope", SpvCapabilityVulkanMemoryModelDeviceScope},
+        {"FPFastMathModeINTEL", SpvCapabilityFPFastMathModeINTEL},
+        {"GroupNonUniformRotateKHR", SpvCapabilityGroupNonUniformRotateKHR},
+        {"DerivativeControl", SpvCapabilityDerivativeControl},
+        {"AtomicStorageOps", SpvCapabilityAtomicStorageOps},
+        {"ImageRect", SpvCapabilityImageRect},
+        {"GroupNonUniformShuffle", SpvCapabilityGroupNonUniformShuffle},
+        {"Linkage", SpvCapabilityLinkage},
+        {"SplitBarrierINTEL", SpvCapabilitySplitBarrierINTEL},
+        {"StorageBufferArrayNonUniformIndexing", SpvCapabilityStorageBufferArrayNonUniformIndexing},
+        {"FPGALoopControlsINTEL", SpvCapabilityFPGALoopControlsINTEL},
+        {"InputAttachment", SpvCapabilityInputAttachment},
+        {"SignedZeroInfNanPreserve", SpvCapabilitySignedZeroInfNanPreserve},
+        {"WorkgroupMemoryExplicitLayout8BitAccessKHR", SpvCapabilityWorkgroupMemoryExplicitLayout8BitAccessKHR},
+        {"StorageBuffer8BitAccess", SpvCapabilityStorageBuffer8BitAccess},
+        {"RayTracingOpacityMicromapEXT", SpvCapabilityRayTracingOpacityMicromapEXT},
+        {"ShaderStereoViewNV", SpvCapabilityShaderStereoViewNV},
+        {"Sampled1D", SpvCapabilitySampled1D},
+        {"Float16Buffer", SpvCapabilityFloat16Buffer},
+        {"FPGAClusterAttributesINTEL", SpvCapabilityFPGAClusterAttributesINTEL},
+        {"NamedBarrier", SpvCapabilityNamedBarrier},
+        {"SubgroupAvcMotionEstimationIntraINTEL", SpvCapabilitySubgroupAvcMotionEstimationIntraINTEL},
+        {"FPGALatencyControlINTEL", SpvCapabilityFPGALatencyControlINTEL},
+        {"ShadingRateNV", SpvCapabilityShadingRateNV},
+        {"StorageUniformBufferBlock16", SpvCapabilityStorageUniformBufferBlock16},
+        {"Float64", SpvCapabilityFloat64},
+        {"CooperativeMatrixNV", SpvCapabilityCooperativeMatrixNV},
+        {"ImageReadWriteLodAMD", SpvCapabilityImageReadWriteLodAMD},
+        {"StorageBufferArrayNonUniformIndexingEXT", SpvCapabilityStorageBufferArrayNonUniformIndexingEXT},
+        {"InterpolationFunction", SpvCapabilityInterpolationFunction},
+        {"FPMaxErrorINTEL", SpvCapabilityFPMaxErrorINTEL},
+        {"PipeStorage", SpvCapabilityPipeStorage},
+        {"FragmentFullyCoveredEXT", SpvCapabilityFragmentFullyCoveredEXT},
+        {"CoreBuiltinsARM", SpvCapabilityCoreBuiltinsARM},
+        {"DeviceEnqueue", SpvCapabilityDeviceEnqueue},
+        {"AtomicFloat32AddEXT", SpvCapabilityAtomicFloat32AddEXT},
+        {"FPGAInvocationPipeliningAttributesINTEL", SpvCapabilityFPGAInvocationPipeliningAttributesINTEL},
+        {"StorageImageArrayNonUniformIndexing", SpvCapabilityStorageImageArrayNonUniformIndexing},
+        {"DotProduct", SpvCapabilityDotProduct},
+        {"RayTracingPositionFetchKHR", SpvCapabilityRayTracingPositionFetchKHR},
+        {"ImageFootprintNV", SpvCapabilityImageFootprintNV},
+        {"TileImageColorReadAccessEXT", SpvCapabilityTileImageColorReadAccessEXT},
+        {"BindlessTextureNV", SpvCapabilityBindlessTextureNV},
+        {"DotProductInputAllKHR", SpvCapabilityDotProductInputAllKHR},
+        {"GroupNonUniformVote", SpvCapabilityGroupNonUniformVote},
+        {"FragmentBarycentricNV", SpvCapabilityFragmentBarycentricNV},
+        {"SampleMaskOverrideCoverageNV", SpvCapabilitySampleMaskOverrideCoverageNV},
+        {"DemoteToHelperInvocationEXT", SpvCapabilityDemoteToHelperInvocationEXT},
+        {"SubgroupBufferBlockIOINTEL", SpvCapabilitySubgroupBufferBlockIOINTEL},
+        {"TextureSampleWeightedQCOM", SpvCapabilityTextureSampleWeightedQCOM},
+        {"SampledImageArrayDynamicIndexing", SpvCapabilitySampledImageArrayDynamicIndexing},
+        {"StorageImageArrayDynamicIndexing", SpvCapabilityStorageImageArrayDynamicIndexing},
+        {"InputAttachmentArrayDynamicIndexing", SpvCapabilityInputAttachmentArrayDynamicIndexing},
+        {"StorageTexelBufferArrayDynamicIndexingEXT", SpvCapabilityStorageTexelBufferArrayDynamicIndexingEXT},
+        {"GroupUniformArithmeticKHR", SpvCapabilityGroupUniformArithmeticKHR},
+        {"FragmentShaderSampleInterlockEXT", SpvCapabilityFragmentShaderSampleInterlockEXT},
+        {"MinLod", SpvCapabilityMinLod},
+        {"GroupNonUniform", SpvCapabilityGroupNonUniform},
+        {"Matrix", SpvCapabilityMatrix},
+        {"VariablePointersStorageBuffer", SpvCapabilityVariablePointersStorageBuffer},
+        {"SampledImageArrayNonUniformIndexing", SpvCapabilitySampledImageArrayNonUniformIndexing},
+        {"StorageImageExtendedFormats", SpvCapabilityStorageImageExtendedFormats},
+        {"VariablePointers", SpvCapabilityVariablePointers},
+        {"RayTracingProvisionalKHR", SpvCapabilityRayTracingProvisionalKHR},
+        {"GeometryShaderPassthroughNV", SpvCapabilityGeometryShaderPassthroughNV},
+        {"DrawParameters", SpvCapabilityDrawParameters},
+        {"TessellationPointSize", SpvCapabilityTessellationPointSize},
+        {"TextureBlockMatchQCOM", SpvCapabilityTextureBlockMatchQCOM},
+        {"WorkgroupMemoryExplicitLayout16BitAccessKHR", SpvCapabilityWorkgroupMemoryExplicitLayout16BitAccessKHR},
+        {"SubgroupImageMediaBlockIOINTEL", SpvCapabilitySubgroupImageMediaBlockIOINTEL},
+        {"GroupNonUniformArithmetic", SpvCapabilityGroupNonUniformArithmetic},
+        {"AtomicStorage", SpvCapabilityAtomicStorage},
+        {"FPGAKernelAttributesINTEL", SpvCapabilityFPGAKernelAttributesINTEL},
+        {"RuntimeDescriptorArray", SpvCapabilityRuntimeDescriptorArray},
+        {"CooperativeMatrixKHR", SpvCapabilityCooperativeMatrixKHR},
+        {"Int64", SpvCapabilityInt64},
+        {"UniformTexelBufferArrayDynamicIndexingEXT", SpvCapabilityUniformTexelBufferArrayDynamicIndexingEXT},
+        {"FragmentShadingRateKHR", SpvCapabilityFragmentShadingRateKHR},
+        {"UniformAndStorageBuffer8BitAccess", SpvCapabilityUniformAndStorageBuffer8BitAccess},
+        {"UniformTexelBufferArrayNonUniformIndexingEXT", SpvCapabilityUniformTexelBufferArrayNonUniformIndexingEXT},
+        {"RoundingModeRTE", SpvCapabilityRoundingModeRTE},
+        {"UniformBufferArrayDynamicIndexing", SpvCapabilityUniformBufferArrayDynamicIndexing},
+        {"Int8", SpvCapabilityInt8},
+        {"VulkanMemoryModelDeviceScopeKHR", SpvCapabilityVulkanMemoryModelDeviceScopeKHR},
+        {"LiteralSampler", SpvCapabilityLiteralSampler},
+        {"IndirectReferencesINTEL", SpvCapabilityIndirectReferencesINTEL},
+        {"USMStorageClassesINTEL", SpvCapabilityUSMStorageClassesINTEL},
+        {"Vector16", SpvCapabilityVector16},
+        {"IOPipesINTEL", SpvCapabilityIOPipesINTEL},
+        {"VectorComputeINTEL", SpvCapabilityVectorComputeINTEL},
+        {"StoragePushConstant16", SpvCapabilityStoragePushConstant16},
+        {"GroupNonUniformShuffleRelative", SpvCapabilityGroupNonUniformShuffleRelative},
+        {"ShaderNonUniform", SpvCapabilityShaderNonUniform},
+        {"CullDistance", SpvCapabilityCullDistance},
+        {"ShaderViewportIndexLayerNV", SpvCapabilityShaderViewportIndexLayerNV},
+        {"GeometryStreams", SpvCapabilityGeometryStreams},
+        {"DotProductInputAll", SpvCapabilityDotProductInputAll},
+        {"GenericPointer", SpvCapabilityGenericPointer},
+        {"DotProductInput4x8BitPackedKHR", SpvCapabilityDotProductInput4x8BitPackedKHR},
+        {"PhysicalStorageBufferAddresses", SpvCapabilityPhysicalStorageBufferAddresses},
+        {"FragmentMaskAMD", SpvCapabilityFragmentMaskAMD},
+        {"SubgroupVoteKHR", SpvCapabilitySubgroupVoteKHR},
+        {"RuntimeAlignedAttributeINTEL", SpvCapabilityRuntimeAlignedAttributeINTEL},
+        {"TileImageStencilReadAccessEXT", SpvCapabilityTileImageStencilReadAccessEXT},
+        {"SubgroupAvcMotionEstimationChromaINTEL", SpvCapabilitySubgroupAvcMotionEstimationChromaINTEL},
+        {"StorageBuffer16BitAccess", SpvCapabilityStorageBuffer16BitAccess},
+        {"RayTracingMotionBlurNV", SpvCapabilityRayTracingMotionBlurNV},
+        {"BFloat16ConversionINTEL", SpvCapabilityBFloat16ConversionINTEL},
+        {"FPGAArgumentInterfacesINTEL", SpvCapabilityFPGAArgumentInterfacesINTEL},
+        {"RayQueryProvisionalKHR", SpvCapabilityRayQueryProvisionalKHR},
+        {"FPGAKernelAttributesv2INTEL", SpvCapabilityFPGAKernelAttributesv2INTEL},
+        {"Addresses", SpvCapabilityAddresses},
+        {"UniformTexelBufferArrayNonUniformIndexing", SpvCapabilityUniformTexelBufferArrayNonUniformIndexing},
+        {"AtomicFloat64MinMaxEXT", SpvCapabilityAtomicFloat64MinMaxEXT},
+        {"SampledBuffer", SpvCapabilitySampledBuffer},
+        {"DebugInfoModuleINTEL", SpvCapabilityDebugInfoModuleINTEL},
+        {"Geometry", SpvCapabilityGeometry},
+        {"LongConstantCompositeINTEL", SpvCapabilityLongConstantCompositeINTEL},
+        {"FragmentShaderPixelInterlockEXT", SpvCapabilityFragmentShaderPixelInterlockEXT},
+        {"VulkanMemoryModelKHR", SpvCapabilityVulkanMemoryModelKHR},
+        {"ShaderViewportIndex", SpvCapabilityShaderViewportIndex},
+        {"DotProductKHR", SpvCapabilityDotProductKHR},
+        {"Groups", SpvCapabilityGroups},
+        {"UniformAndStorageBuffer16BitAccess", SpvCapabilityUniformAndStorageBuffer16BitAccess},
+        {"VectorAnyINTEL", SpvCapabilityVectorAnyINTEL},
+        {"InputAttachmentArrayDynamicIndexingEXT", SpvCapabilityInputAttachmentArrayDynamicIndexingEXT},
+        {"Float16ImageAMD", SpvCapabilityFloat16ImageAMD},
+        {"ShaderViewportIndexLayerEXT", SpvCapabilityShaderViewportIndexLayerEXT},
+        {"ImageGatherExtended", SpvCapabilityImageGatherExtended},
+        {"SampleMaskPostDepthCoverage", SpvCapabilitySampleMaskPostDepthCoverage},
+        {"FragmentBarycentricKHR", SpvCapabilityFragmentBarycentricKHR},
+        {"TextureBoxFilterQCOM", SpvCapabilityTextureBoxFilterQCOM},
+        {"ShaderSMBuiltinsNV", SpvCapabilityShaderSMBuiltinsNV},
+        {"ImageReadWrite", SpvCapabilityImageReadWrite},
+        {"StorageImageMultisample", SpvCapabilityStorageImageMultisample},
+        {"ArbitraryPrecisionFixedPointINTEL", SpvCapabilityArbitraryPrecisionFixedPointINTEL},
+        {"ComputeDerivativeGroupQuadsNV", SpvCapabilityComputeDerivativeGroupQuadsNV},
+        {"IntegerFunctions2INTEL", SpvCapabilityIntegerFunctions2INTEL},
+        {"AtomicFloat32MinMaxEXT", SpvCapabilityAtomicFloat32MinMaxEXT},
+        {"BitInstructions", SpvCapabilityBitInstructions},
+        {"AtomicFloat16MinMaxEXT", SpvCapabilityAtomicFloat16MinMaxEXT},
+        {"SubgroupShuffleINTEL", SpvCapabilitySubgroupShuffleINTEL},
+        {"GroupNonUniformPartitionedNV", SpvCapabilityGroupNonUniformPartitionedNV},
+        {"SampledRect", SpvCapabilitySampledRect},
+        {"GroupNonUniformClustered", SpvCapabilityGroupNonUniformClustered},
+        {"StorageImageReadWithoutFormat", SpvCapabilityStorageImageReadWithoutFormat},
+        {"ImageBasic", SpvCapabilityImageBasic},
+        {"FPGARegINTEL", SpvCapabilityFPGARegINTEL},
+        {"SparseResidency", SpvCapabilitySparseResidency},
+        {"ImageGatherBiasLodAMD", SpvCapabilityImageGatherBiasLodAMD},
+        {"OptNoneINTEL", SpvCapabilityOptNoneINTEL},
+        {"ShaderLayer", SpvCapabilityShaderLayer},
+        {"StorageImageWriteWithoutFormat", SpvCapabilityStorageImageWriteWithoutFormat},
+        {"Float16", SpvCapabilityFloat16},
+        {"FragmentShaderShadingRateInterlockEXT", SpvCapabilityFragmentShaderShadingRateInterlockEXT},
+        {"SubgroupImageBlockIOINTEL", SpvCapabilitySubgroupImageBlockIOINTEL},
+        {"ShaderInvocationReorderNV", SpvCapabilityShaderInvocationReorderNV},
+        {"PhysicalStorageBufferAddressesEXT", SpvCapabilityPhysicalStorageBufferAddressesEXT},
+        {"UniformBufferArrayNonUniformIndexingEXT", SpvCapabilityUniformBufferArrayNonUniformIndexingEXT},
+        {"ShaderViewportMaskNV", SpvCapabilityShaderViewportMaskNV},
+        {"DenormPreserve", SpvCapabilityDenormPreserve},
+        {"FPGADSPControlINTEL", SpvCapabilityFPGADSPControlINTEL},
+        {"SampleRateShading", SpvCapabilitySampleRateShading},
+        {"RoundingModeRTZ", SpvCapabilityRoundingModeRTZ},
+        {"Kernel", SpvCapabilityKernel},
+        {"FunctionPointersINTEL", SpvCapabilityFunctionPointersINTEL},
+        {"FloatingPointModeINTEL", SpvCapabilityFloatingPointModeINTEL},
+        {"StorageInputOutput16", SpvCapabilityStorageInputOutput16},
+        {"AtomicFloat16AddEXT", SpvCapabilityAtomicFloat16AddEXT},
+        {"ImageCubeArray", SpvCapabilityImageCubeArray},
+        {"DotProductInput4x8BitPacked", SpvCapabilityDotProductInput4x8BitPacked},
+        {"ClipDistance", SpvCapabilityClipDistance},
+        {"Int64Atomics", SpvCapabilityInt64Atomics},
+        {"MultiViewport", SpvCapabilityMultiViewport},
+        {"RayQueryPositionFetchKHR", SpvCapabilityRayQueryPositionFetchKHR},
+        {"StencilExportEXT", SpvCapabilityStencilExportEXT},
+        {"RuntimeDescriptorArrayEXT", SpvCapabilityRuntimeDescriptorArrayEXT},
+        {"RoundToInfinityINTEL", SpvCapabilityRoundToInfinityINTEL},
+        {"StorageTexelBufferArrayDynamicIndexing", SpvCapabilityStorageTexelBufferArrayDynamicIndexing},
+        {"DotProductInput4x8Bit", SpvCapabilityDotProductInput4x8Bit},
+        {"UnstructuredLoopControlsINTEL", SpvCapabilityUnstructuredLoopControlsINTEL},
+        {"WorkgroupMemoryExplicitLayoutKHR", SpvCapabilityWorkgroupMemoryExplicitLayoutKHR},
+        {"TransformFeedback", SpvCapabilityTransformFeedback},
+        {"Tessellation", SpvCapabilityTessellation},
+        {"SubgroupBallotKHR", SpvCapabilitySubgroupBallotKHR},
+        {"ImageMipmap", SpvCapabilityImageMipmap},
+        {"DeviceGroup", SpvCapabilityDeviceGroup},
+        {"Pipes", SpvCapabilityPipes},
+        {"FunctionFloatControlINTEL", SpvCapabilityFunctionFloatControlINTEL},
+        {"RayCullMaskKHR", SpvCapabilityRayCullMaskKHR},
+        {"SampledImageArrayNonUniformIndexingEXT", SpvCapabilitySampledImageArrayNonUniformIndexingEXT},
+        {"DemoteToHelperInvocation", SpvCapabilityDemoteToHelperInvocation},
+        {"MeshShadingEXT", SpvCapabilityMeshShadingEXT},
+        {"RayTraversalPrimitiveCullingKHR", SpvCapabilityRayTraversalPrimitiveCullingKHR},
+        {"GroupNonUniformQuad", SpvCapabilityGroupNonUniformQuad},
+        {"FPGAMemoryAttributesINTEL", SpvCapabilityFPGAMemoryAttributesINTEL},
+        {"GeometryPointSize", SpvCapabilityGeometryPointSize},
+        {"FPGAMemoryAccessesINTEL", SpvCapabilityFPGAMemoryAccessesINTEL},
+        {"Int16", SpvCapabilityInt16},
+        {"GroupNonUniformBallot", SpvCapabilityGroupNonUniformBallot},
+        {"Int64ImageEXT", SpvCapabilityInt64ImageEXT},
+        {"PerViewAttributesNV", SpvCapabilityPerViewAttributesNV},
+        {"ArbitraryPrecisionFloatingPointINTEL", SpvCapabilityArbitraryPrecisionFloatingPointINTEL},
+        {"KernelAttributesINTEL", SpvCapabilityKernelAttributesINTEL},
+        {"AtomicFloat64AddEXT", SpvCapabilityAtomicFloat64AddEXT},
+        {"FPGABufferLocationINTEL", SpvCapabilityFPGABufferLocationINTEL},
+        {"StorageTexelBufferArrayNonUniformIndexing", SpvCapabilityStorageTexelBufferArrayNonUniformIndexing},
+        {"ImageBuffer", SpvCapabilityImageBuffer},
+        {"TileImageDepthReadAccessEXT", SpvCapabilityTileImageDepthReadAccessEXT},
+        {"ImageMSArray", SpvCapabilityImageMSArray},
+        {"StorageTexelBufferArrayNonUniformIndexingEXT", SpvCapabilityStorageTexelBufferArrayNonUniformIndexingEXT},
+        {"MultiView", SpvCapabilityMultiView},
+        {"InputAttachmentArrayNonUniformIndexingEXT", SpvCapabilityInputAttachmentArrayNonUniformIndexingEXT},
+        {"ExpectAssumeKHR", SpvCapabilityExpectAssumeKHR},
+        {"RayQueryKHR", SpvCapabilityRayQueryKHR},
+        {"BlockingPipesINTEL", SpvCapabilityBlockingPipesINTEL},
+        {"StorageBufferArrayDynamicIndexing", SpvCapabilityStorageBufferArrayDynamicIndexing},
+        {"MemoryAccessAliasingINTEL", SpvCapabilityMemoryAccessAliasingINTEL},
+        {"MeshShadingNV", SpvCapabilityMeshShadingNV},
+        {"RayTracingKHR", SpvCapabilityRayTracingKHR},
+        {"RayTracingNV", SpvCapabilityRayTracingNV},
+        {"StorageUniform16", SpvCapabilityStorageUniform16},
+        {"DotProductInput4x8BitKHR", SpvCapabilityDotProductInput4x8BitKHR},
+        {"StoragePushConstant8", SpvCapabilityStoragePushConstant8},
+        {"VariableLengthArrayINTEL", SpvCapabilityVariableLengthArrayINTEL},
+        {"ShaderNonUniformEXT", SpvCapabilityShaderNonUniformEXT},
+        {"SubgroupAvcMotionEstimationINTEL", SpvCapabilitySubgroupAvcMotionEstimationINTEL},
+        {"StorageImageArrayNonUniformIndexingEXT", SpvCapabilityStorageImageArrayNonUniformIndexingEXT},
+        {"FragmentDensityEXT", SpvCapabilityFragmentDensityEXT},
+        {"LoopFuseINTEL", SpvCapabilityLoopFuseINTEL},
+        {"ImageQuery", SpvCapabilityImageQuery},
+    };
+
+    static const auto hash = [](const UnownedStringSlice& str, UInt32 salt){
+        return combineHash(getHashCode(str), getHashCode(salt)) % 245;
+    };
+
+    const auto i = hash(str, tableSalt[hash(str, 0)]);
+    if(str == words[i].first)
+    {
+        value = words[i].second;
+        return true;
+    }
+    else
+    {
+        return false;
+    }
+}
+
 RefPtr<SPIRVCoreGrammarInfo> getEmbeddedSPIRVCoreGrammarInfo()
 {
     static SPIRVCoreGrammarInfo info = [](){
         SPIRVCoreGrammarInfo info;
         info.spvOps.embedded = &lookupSpvOp;
+        info.spvCapabilities.embedded = &lookupSpvCapability;
         info.addReference();
         return info;
     }();
